+++ src/renderer/shared/ui-kit/index.ts
@@ -1,6 +1,7 @@
 export { useResizeObserver } from './hooks/useResizeObserver';
 export { useTheme } from './Theme/useTheme';
 
+export { EmptyMessage } from './EmptyMessage/EmptyMessage';
 export { SearchInput } from './SearchInput/SearchInput';
 export { ThemeProvider } from './Theme/ThemeProvider';
 export { ScrollArea } from './ScrollArea/ScrollArea';
